feat: implent basic life cycle hooks
