+++ packages/core/src/context.ts
@@ -1,0 +1,5 @@
+export const ctx = {
+  help() {
+    console.log('HELP!')
+  },
+}
