+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
-export * from './run'
+export { runCLI } from './run'
+export { defineCommand } from './command'
+export { defineFlag } from './flag'
